requirements.txt
django-cors-headers==3.5.0
asgiref==3.2.7
boto3==1.13.9
botocore==1.16.9
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
google-api-core==1.17.0
google-api-python-client==1.8.3

google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
jmespath==0.10.0
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
